requirements.txt
pipenv==11.9.0
python-dotenv==0.20.0
python-telegram-bot==13.13
requests==2.25.1

